+++ pandas/tseries/tests/test_timezones.py
@@ -363,7 +363,7 @@ class TestTimeZoneSupportPytz(tm.TestCase):
         dr = date_range('2011-10-02 00:00', freq='h', periods=10,
                         tz=self.tzstr('America/Atikokan'))
 
-        expected = np.arange(10)
+        expected = np.arange(10, dtype=np.int32)
         self.assert_numpy_array_equal(dr.hour, expected)
 
     def test_with_tz(self):
@@ -907,7 +907,8 @@ class TestTimeZoneSupportDateutil(TestTimeZoneSupportPytz):
             idx = idx.tz_localize('UTC')
             idx = idx.tz_convert('Europe/Moscow')
 
-            expected = np.repeat(np.array([3, 4, 5]), np.array([n, n, 1]))
+            expected = np.repeat(np.array([3, 4, 5], dtype=np.int32),
+                                 np.array([n, n, 1]))
             self.assert_numpy_array_equal(idx.hour, expected)
 
     def test_tslib_tz_convert_dst(self):
@@ -917,14 +918,15 @@ class TestTimeZoneSupportDateutil(TestTimeZoneSupportPytz):
                              tz='UTC')
             idx = idx.tz_convert('US/Eastern')
             expected = np.repeat(np.array([18, 19, 20, 21, 22, 23,
-                                           0, 1, 3, 4, 5]),
+                                           0, 1, 3, 4, 5], dtype=np.int32),
                                  np.array([n, n, n, n, n, n, n, n, n, n, 1]))
             self.assert_numpy_array_equal(idx.hour, expected)
 
             idx = date_range('2014-03-08 18:00', '2014-03-09 05:00', freq=freq,
                              tz='US/Eastern')
             idx = idx.tz_convert('UTC')
-            expected = np.repeat(np.array([23, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9]),
+            expected = np.repeat(np.array([23, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9],
+                                          dtype=np.int32),
                                  np.array([n, n, n, n, n, n, n, n, n, n, 1]))
             self.assert_numpy_array_equal(idx.hour, expected)
 
@@ -933,7 +935,7 @@ class TestTimeZoneSupportDateutil(TestTimeZoneSupportPytz):
                              tz='UTC')
             idx = idx.tz_convert('US/Eastern')
             expected = np.repeat(np.array([19, 20, 21, 22, 23,
-                                           0, 1, 1, 2, 3, 4]),
+                                           0, 1, 1, 2, 3, 4], dtype=np.int32),
                                  np.array([n, n, n, n, n, n, n, n, n, n, 1]))
             self.assert_numpy_array_equal(idx.hour, expected)
 
@@ -941,7 +943,7 @@ class TestTimeZoneSupportDateutil(TestTimeZoneSupportPytz):
                              tz='US/Eastern')
             idx = idx.tz_convert('UTC')
             expected = np.repeat(np.array([22, 23, 0, 1, 2, 3, 4, 5, 6,
-                                           7, 8, 9, 10]),
+                                           7, 8, 9, 10], dtype=np.int32),
                                  np.array([n, n, n, n, n, n, n, n, n,
                                            n, n, n, 1]))
             self.assert_numpy_array_equal(idx.hour, expected)
@@ -951,23 +953,27 @@ class TestTimeZoneSupportDateutil(TestTimeZoneSupportPytz):
         idx = date_range('2014-03-08 00:00', '2014-03-09 00:00', freq='D',
                          tz='UTC')
         idx = idx.tz_convert('US/Eastern')
-        self.assert_numpy_array_equal(idx.hour, np.array([19, 19]))
+        self.assert_numpy_array_equal(idx.hour,
+                                      np.array([19, 19], dtype=np.int32))
 
         idx = date_range('2014-03-08 00:00', '2014-03-09 00:00', freq='D',
                          tz='US/Eastern')
         idx = idx.tz_convert('UTC')
-        self.assert_numpy_array_equal(idx.hour, np.array([5, 5]))
+        self.assert_numpy_array_equal(idx.hour,
+                                      np.array([5, 5], dtype=np.int32))
 
         # End DST
         idx = date_range('2014-11-01 00:00', '2014-11-02 00:00', freq='D',
                          tz='UTC')
         idx = idx.tz_convert('US/Eastern')
-        self.assert_numpy_array_equal(idx.hour, np.array([20, 20]))
+        self.assert_numpy_array_equal(idx.hour,
+                                      np.array([20, 20], dtype=np.int32))
 
         idx = date_range('2014-11-01 00:00', '2014-11-02 000:00', freq='D',
                          tz='US/Eastern')
         idx = idx.tz_convert('UTC')
-        self.assert_numpy_array_equal(idx.hour, np.array([4, 4]))
+        self.assert_numpy_array_equal(idx.hour,
+                                      np.array([4, 4], dtype=np.int32))
 
 
 class TestTimeZoneCacheKey(tm.TestCase):
